adding delete and update saga - needs refactoring

+++ services/authentication/src/main/java/com/graduationProject/authentication/saga/sagaParticipators/CreateResident.java
@@ -3,7 +3,6 @@ package com.graduationProject.authentication.saga.sagaParticipators;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.graduationProject.authentication.dto.ResidentDto;
 import com.graduationProject.authentication.dto.saga.SagaResidentDto;
 import com.graduationProject.authentication.dto.saga.SagaResponseDto;
 import com.graduationProject.authentication.kafka.KafkaApi;
@@ -13,8 +12,10 @@ import com.graduationProject.authentication.type.SagaStatus;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import static com.graduationProject.authentication.topic.ResidentTopics.CreateResidentSagaDone;
-import static com.graduationProject.authentication.topic.ResidentTopics.CreateResidentSagaRevert;
+import javax.transaction.Transactional;
+
+import static com.graduationProject.authentication.topic.ResidentTopic.CreateResidentSagaDone;
+import static com.graduationProject.authentication.topic.ResidentTopic.CreateResidentSagaRevert;
 
 
 @Service
@@ -30,18 +31,19 @@ public class CreateResident implements SagaParticipator<SagaResidentDto> {
         this.kafkaApi = kafkaApi;
     }
 
+    @Transactional
     @Override
     public void transact(SagaResidentDto sagaResidentDto) {
         //if anything goes wrong it should publish to the revert topic
         try {
-            residentService.addResident(sagaResidentDto.getResidentDto());
+
             SagaResponseDto sagaResponseDto = new SagaResponseDto(sagaResidentDto.getSagaId(),
                     SagaStatus.SUCCESS);
             if (sagaResidentDto.getResidentDto().getUsername().equals("fail")) {
                 throw new Exception(); //this should trigger a revert of the saga.
             }
+            residentService.addResident(sagaResidentDto.getResidentDto());
             kafkaApi.publish(CreateResidentSagaDone, new ObjectMapper().writeValueAsString(sagaResponseDto));
-            System.out.println("to topic " + CreateResidentSagaDone + " i sent this: " + sagaResponseDto.toString());
         } catch (Exception e) {
             SagaResponseDto sagaResponseDto = new SagaResponseDto(sagaResidentDto.getSagaId(),
                     SagaStatus.FAILED);
@@ -55,6 +57,7 @@ public class CreateResident implements SagaParticipator<SagaResidentDto> {
 
     }
 
+    @Transactional
     @Override
     public void revert(SagaResidentDto sagaResidentDto) {
         /*
@@ -70,7 +73,10 @@ public class CreateResident implements SagaParticipator<SagaResidentDto> {
                 throw new Exception();
             }
 
-            residentService.deleteResident(sagaResidentDto.getResidentDto().getId());
+            if (residentService.residentExists(sagaResidentDto.getResidentDto().getId())) {
+                residentService.deleteResident(sagaResidentDto.getResidentDto().getId());
+            }
+
             kafkaApi.publish(CreateResidentSagaRevert, new ObjectMapper()
                     .writeValueAsString(new SagaResponseDto(sagaResidentDto.getSagaId(), SagaStatus.SUCCESS)));
         } catch (Exception e) {
